requirements.txt
asgiref==3.8.1
certifi==2025.4.26
crispy-bootstrap5==2025.4
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
djangorestframework==3.16.0
filelock==3.18.0
packaging==25.0
pipenv==2025.0.2
platformdirs==4.3.8
psycopg2-binary==2.9.10
setuptools==80.8.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.31.2
